+++ MathForGames/Goal.cs
@@ -40,7 +40,7 @@ namespace MathForGames
         /// <returns></returns>
         private bool CheckPlayerDistance()
         {
-            float distance = (_player.Position - Position).Magnitude;
+            double distance = (_player.Position - Position).Magnitude;
             return distance <= 1;
         }
